basic_a2a/pyproject.toml
[project]
name = "geo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
